+++ ssr/feature/api-v1/routes/v1-document-play.test.e2e.js
@@ -11,7 +11,7 @@ const getQueueMetrics = async (queue) => {
 };
 
 describe('v1QueueDocumentPlay', () => {
-  beforeEach(global.dropAllQueues)
+  beforeEach(global.dropAllQueues);
 
   it('should play an existing document', async () => {
     await global.fetchq.query(`SELECT * FROM fetchq.queue_create('q1')`);
@@ -58,7 +58,9 @@ describe('v1QueueDocumentPlay', () => {
 
   it('should update counters from ACTIVE status', async () => {
     await global.fetchq.query(`SELECT FROM fetchq.queue_create('q1')`);
-    await global.fetchq.query(`SELECT FROM fetchq.queue_set_max_attempts('q1', 1)`);
+    await global.fetchq.query(
+      `SELECT FROM fetchq.queue_set_max_attempts('q1', 1)`,
+    );
     await global.fetchq.query(`SELECT FROM fetchq.doc_push('q1', 's1')`);
     await global.fetchq.query(`SELECT FROM fetchq.doc_pick('q1', 0, 1, '1y')`);
     await global.fetchq.query(`SELECT FROM fetchq.mnt()`);
@@ -75,7 +77,9 @@ describe('v1QueueDocumentPlay', () => {
 
   it('should update counters from PLANNED status', async () => {
     await global.fetchq.query(`SELECT FROM fetchq.queue_create('q1')`);
-    await global.fetchq.query(`SELECT FROM fetchq.queue_set_max_attempts('q1', 1)`);
+    await global.fetchq.query(
+      `SELECT FROM fetchq.queue_set_max_attempts('q1', 1)`,
+    );
     await global.fetchq.query(
       `SELECT FROM fetchq.doc_push('q1', 's1', 0, 0, NOW() + INTERVAL '1m', '{}')`,
     );
@@ -93,7 +97,9 @@ describe('v1QueueDocumentPlay', () => {
 
   it('should update counters from COMPLETED status', async () => {
     await global.fetchq.query(`SELECT FROM fetchq.queue_create('q1')`);
-    await global.fetchq.query(`SELECT FROM fetchq.queue_set_max_attempts('q1', 1)`);
+    await global.fetchq.query(
+      `SELECT FROM fetchq.queue_set_max_attempts('q1', 1)`,
+    );
     await global.fetchq.query(`SELECT FROM fetchq.doc_push('q1', 's1')`);
     await global.fetchq.query(`SELECT FROM fetchq.doc_pick('q1', 0, 1, '1y')`);
     await global.fetchq.query(`SELECT FROM fetchq.doc_complete('q1', 's1')`);
@@ -129,7 +135,9 @@ describe('v1QueueDocumentPlay', () => {
 
   it('should update counters from KILLED status', async () => {
     await global.fetchq.query(`SELECT FROM fetchq.queue_create('q1')`);
-    await global.fetchq.query(`SELECT FROM fetchq.queue_set_max_attempts('q1', 1)`);
+    await global.fetchq.query(
+      `SELECT FROM fetchq.queue_set_max_attempts('q1', 1)`,
+    );
     await global.fetchq.query(`SELECT FROM fetchq.doc_push('q1', 's1')`);
     await global.fetchq.query(`SELECT FROM fetchq.doc_pick('q1', 0, 1, '1y')`);
     await global.fetchq.query(`SELECT FROM fetchq.doc_kill('q1', 's1')`);