+++ ssr/feature/api-v1/routes/v1-queue-details.test.e2e.js
@@ -1,5 +1,5 @@
 describe('v1QueueDetails', () => {
-  beforeEach(global.dropAllQueues)
+  beforeEach(global.dropAllQueues);
 
   const getMetric = (metric) => ($) => $.metric === metric;