+++ ssr/feature/auth-v1/decorators/authenticate.js
@@ -39,7 +39,7 @@ const authenticateDecorator = async (request, reply) => {
 
   // Validate the token
   try {
-    const { payload } = await request.jwt.verify(token);
+    const { payload } = await request.jwt.verify(token);
     request.auth = payload;
   } catch (err) {
     return reply.send({