+++ ssr/feature/auth-v1/routes/v1.session-details.test.e2e.js
@@ -1,21 +1,25 @@
-
 describe('v1/session/details', () => {
   const mockPassword = async (value = null) => {
     const PWD_PATH = 'app.auth.console.password';
-    const { value: currentPassword} = await global.get(`/test/config?key=${PWD_PATH}`)
-    await global.post(`/test/config`, { key: PWD_PATH, value})
+    const { value: currentPassword } = await global.get(
+      `/test/config?key=${PWD_PATH}`,
+    );
+    await global.post(`/test/config`, { key: PWD_PATH, value });
 
-    return mockPassword.resetMock = () => {
-      mockPassword.resetMock = null
-      return global.post(`/test/config`, { key: PWD_PATH, value: currentPassword})
-    }
-  }
+    return (mockPassword.resetMock = () => {
+      mockPassword.resetMock = null;
+      return global.post(`/test/config`, {
+        key: PWD_PATH,
+        value: currentPassword,
+      });
+    });
+  };
 
-  afterEach(async() => {
+  afterEach(async () => {
     if (mockPassword.resetMock) {
-      await mockPassword.resetMock()
+      await mockPassword.resetMock();
     }
-  })
+  });
 
   it('should receive a dummy session for a non secured instance', async () => {
     await mockPassword(null);
@@ -35,14 +39,16 @@ describe('v1/session/details', () => {
     await mockPassword('foobar');
 
     // Obtain a valid session token by authentication method
-    const { value: headerName } = await global.get(`/test/config?key=app.auth.header.name`)
+    const { value: headerName } = await global.get(
+      `/test/config?key=app.auth.header.name`,
+    );
     const r1 = await global.post(`/api/v1/session`, {
       uname: 'console',
       passw: 'foobar',
     });
-    
+
     // Check the session status by forwarding the session token as a Header
-    const headerValue = `Bearer ${r1.data.token}`
+    const headerValue = `Bearer ${r1.data.token}`;
     const r2 = await global.get(`/api/v1/session?q=2`, {
       headers: {
         [headerName]: headerValue,
@@ -57,7 +63,9 @@ describe('v1/session/details', () => {
     await mockPassword('foobar');
 
     // Obtain a valid session token by authentication method
-    const { value: cookieName } = await global.get(`/test/config?key=app.auth.cookie.name`)
+    const { value: cookieName } = await global.get(
+      `/test/config?key=app.auth.cookie.name`,
+    );
 
     const r1 = await global.post(`/api/v1/session`, {
       uname: 'console',
@@ -65,7 +73,7 @@ describe('v1/session/details', () => {
     });
 
     // Check the session status by forwarding the session token as a Cookie
-    const cookieValue = `${cookieName}=${r1.data.token}`
+    const cookieValue = `${cookieName}=${r1.data.token}`;
     const r2 = await global.get(`/api/v1/session?q=2`, {
       headers: {
         Cookie: cookieValue,
@@ -80,7 +88,9 @@ describe('v1/session/details', () => {
     await mockPassword('foobar');
 
     // Obtain a valid session token by authentication method
-    const { value: queryParamName } = await global.get(`/test/config?key=app.auth.query.param`)
+    const { value: queryParamName } = await global.get(
+      `/test/config?key=app.auth.query.param`,
+    );
 
     const r1 = await global.post(`/api/v1/session`, {
       uname: 'console',
@@ -89,9 +99,7 @@ describe('v1/session/details', () => {
 
     // Check the session status by forwarding the session token as a Query
     const queryParam = `${queryParamName}=${r1.data.token}`;
-    const r2 = await global.get(
-      `/api/v1/session?${queryParam}`,
-    );
+    const r2 = await global.get(`/api/v1/session?${queryParam}`);
 
     expect(r2.success).toBe(true);
     expect(r2.data.groups[0]).toBe('*');
@@ -100,7 +108,9 @@ describe('v1/session/details', () => {
   it('should fail to validate a session with a wrong signature', async () => {
     await mockPassword('foobar');
 
-    const { value: headerName } = await global.get(`/test/config?key=app.auth.header.name`)
+    const { value: headerName } = await global.get(
+      `/test/config?key=app.auth.header.name`,
+    );
     const r1 = await global.get(`/api/v1/session?q=2`, {
       headers: {
         [headerName]: `Bearer eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJncm91cHMiOlsiKiJdLCJpYXQiOjE1ODM3NDI1NDUsImV4cCI6MTU4Mzc0MjU0N30.D3E7MPe_uB7TrI-gSwh1Ij_8mefX17AjeRqQ434K7yI`,
@@ -115,7 +125,9 @@ describe('v1/session/details', () => {
   it('should fail to validate a session with an expired token', async () => {
     await mockPassword('foobar');
 
-    const { value: headerName } = await global.get(`/test/config?key=app.auth.header.name`)
+    const { value: headerName } = await global.get(
+      `/test/config?key=app.auth.header.name`,
+    );
     const r1 = await global.get(`/api/v1/session?q=2`, {
       headers: {
         [headerName]: `Bearer eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJncm91cHMiOlsiKiJdLCJpYXQiOjE1ODM3NDI1NDUsImV4cCI6MTU4Mzc0MjU0N30.a1lfsngErvuS_zrY_iBboIcivcrU-UOGkxb9PwphKcs`,